requirements.txt
click>=8.1.7
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
pydantic>=2.8.2
openai>=1.43.0
tenacity>=8.2.3
